Crud restaurantes 0.1

+++ app/src/main/java/com/example/logincardview/domain/usecase/restaurant/DeleteRestaurantUseCase.kt
@@ -3,7 +3,7 @@ package com.example.logincardview.domain.usecase.restaurant
 import com.example.logincardview.data.repository.RestaurantInMemoryRepository
 
 class DeleteRestaurantUseCase(private val repository: RestaurantInMemoryRepository) {
-    suspend operator fun invoke(restaurantId: Int) {
+    suspend operator fun invoke(restaurantId: Long) {
         repository.delete(restaurantId)
     }
 }
